Modularizado el codigo: * Separado el codigo de c en cuatro archivos * Generados headers para los archivos * Generado y escrito archivo makefile para coompilar facilmente * Actualizado gitignore
